--- src/main/java/com/simso/controller/api/PostsApiController.java
@@ -1,29 +1,0 @@
-package com.simso.controller.api;
-
-import com.simso.dto.post.PostsResponseDto;
-import com.simso.dto.post.PostsSaveRequestDto;
-import com.simso.dto.post.PostsUpdateRequestDto;
-import com.simso.service.post.PostService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-@RequiredArgsConstructor
-@RestController
-public class PostsApiController {
-    private final PostService postService;
-
-    @PostMapping("api/v1/posts/")
-    public Long save(@RequestBody PostsSaveRequestDto requestDto){
-        return postService.save(requestDto);
-    }
-
-    @PutMapping("/api/v1/posts/{id}")
-    public Long update(@PathVariable Long id, @RequestBody PostsUpdateRequestDto requestDto){
-        return postService.update(id,requestDto);
-    }
-
-    @GetMapping("/api/v1/posts/{id}")
-    public PostsResponseDto findById (@PathVariable Long id){
-        return postService.findById(id);
-    }
-}
