--- src/main/java/com/simso/controller/homeController.java
@@ -1,25 +1,0 @@
-package com.simso.controller;
-
-import com.simso.config.auth.LoginUser;
-import com.simso.config.auth.dto.SessionUser;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-@RequiredArgsConstructor
-public class homeController {
-
-
-    @GetMapping("/")
-    public String home(Model model, @LoginUser SessionUser user) {
-
-        if (user != null) {
-            model.addAttribute("userName", user.getName());
-        }
-        return "/home";
-    }
-
-
-}
