+++ src/main/java/com/simso/domain/post/dto/PostsListResponseDto.java
@@ -1,0 +1,21 @@
+package com.simso.domain.post.dto;
+
+import com.simso.domain.post.entity.Posts;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class PostsListResponseDto {
+    private Long id;
+    private String title;
+    private String author;
+    private LocalDateTime modifiedDate;
+
+    public PostsListResponseDto(Posts entity) {
+        this.id = entity.getId();
+        this.title = entity.getTitle();
+        this.author = entity.getAuthor();
+        this.modifiedDate = entity.getModifiedDate();
+    }
+}
