+++ src/main/java/com/simso/domain/post/dto/PostsResponseDto.java
@@ -1,0 +1,20 @@
+package com.simso.domain.post.dto;
+
+import com.simso.domain.post.entity.Posts;
+import lombok.Getter;
+
+@Getter
+public class PostsResponseDto {
+
+    private Long id;
+    private String title;
+    private String content;
+    private String author;
+
+    public PostsResponseDto(Posts entity) {
+        this.id = entity.getId();
+        this.title = entity.getTitle();
+        this.content = entity.getContent();
+        this.author = entity.getAuthor();
+    }
+}
