+++ src/main/java/com/simso/domain/post/dto/PostsUpdateRequestDto.java
@@ -1,0 +1,19 @@
+package com.simso.domain.post.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class PostsUpdateRequestDto {
+    private String title;
+    private String content;
+
+
+    @Builder
+    public PostsUpdateRequestDto(String title, String content){
+        this.title = title;
+        this.content = content;
+    }
+}
