+++ src/main/java/com/simso/domain/post/entity/Posts.java
@@ -1,0 +1,44 @@
+package com.simso.domain.post.entity;
+
+import com.simso.global.entity.BaseTimeEntity;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
+@Getter
+@NoArgsConstructor
+@Entity
+public class Posts extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name="title")
+    @NotNull
+    private String title;
+
+    @Column(name = "content", columnDefinition = "TEXT")
+    @NotNull
+    private String content;
+
+    @Column(name = "author")
+    @NotNull
+    private String author;
+
+    @Builder
+    public Posts(Long id, String title, String content, String author) {
+        this.id = id;
+        this.title = title;
+        this.content = content;
+        this.author = author;
+    }
+
+    public void update(String title, String content) {
+        this.title = title;
+        this.content = content;
+    }
+}
