+++ src/main/java/com/simso/domain/post/repository/PostsRepository.java
@@ -1,0 +1,13 @@
+package com.simso.domain.post.repository;
+
+import com.simso.domain.post.entity.Posts;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
+
+public interface PostsRepository extends JpaRepository <Posts, Long> {
+
+    @Query("SELECT p FROM Posts p ORDER BY p.id DESC")
+    List<Posts> findAllDesc();
+}
