+++ src/main/java/com/simso/domain/user/dto/UserSaveRequestDto.java
@@ -1,0 +1,26 @@
+package com.simso.domain.user.dto;
+
+import com.simso.domain.user.entity.User;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class UserSaveRequestDto {
+    private String username;
+    private String password;
+
+    @Builder
+    public UserSaveRequestDto(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public User toEntity() {
+        return User.builder()
+                .username(username)
+                .password(password)
+                .build();
+    }
+}
