+++ src/main/java/com/simso/domain/user/dto/UserUpdateRequestDto.java
@@ -1,0 +1,19 @@
+package com.simso.domain.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class UserUpdateRequestDto {
+    private String username;
+    private String password;
+
+
+    @Builder
+    public UserUpdateRequestDto(String username, String password){
+        this.username = username;
+        this.password = password;
+    }
+}
