+++ src/main/java/com/simso/domain/user/entity/Role.java
@@ -1,0 +1,16 @@
+package com.simso.domain.user.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+
+    GEST("ROLE_GUSET", "손님"),
+    USER("ROLE_USER", "일반 사용자");
+
+    private final String key;
+    private final String title;
+
+}
