+++ src/main/java/com/simso/domain/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.simso.domain.user.repository;
+
+import com.simso.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+   Optional<User> findByUsername(String username);
+
+   Optional<User> findByEmail(String email);
+}
