--- src/main/java/com/simso/dto/PostsListResponseDto.java
@@ -1,21 +1,0 @@
-package com.simso.dto;
-
-import com.simso.domain.Posts;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class PostsListResponseDto {
-    private Long id;
-    private String title;
-    private String author;
-    private LocalDateTime modifiedDate;
-
-    public PostsListResponseDto(Posts entity) {
-        this.id = entity.getId();
-        this.title = entity.getTitle();
-        this.author = entity.getAuthor();
-        this.modifiedDate = entity.getModifiedDate();
-    }
-}
