--- src/main/java/com/simso/dto/UserUpdateRequestDto.java
@@ -1,19 +1,0 @@
-package com.simso.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UserUpdateRequestDto {
-    private String username;
-    private String password;
-
-
-    @Builder
-    public UserUpdateRequestDto(String username, String password){
-        this.username = username;
-        this.password = password;
-    }
-}
