--- src/main/java/com/simso/dto/post/PostsResponseDto.java
@@ -1,20 +1,0 @@
-package com.simso.dto.post;
-
-import com.simso.domain.Posts;
-import lombok.Getter;
-
-@Getter
-public class PostsResponseDto {
-
-    private Long id;
-    private String title;
-    private String content;
-    private String author;
-
-    public PostsResponseDto(Posts entity) {
-        this.id = entity.getId();
-        this.title = entity.getTitle();
-        this.content = entity.getContent();
-        this.author = entity.getAuthor();
-    }
-}
