--- src/main/java/com/simso/dto/post/PostsSaveRequestDto.java
@@ -1,29 +1,0 @@
-package com.simso.dto.post;
-
-import com.simso.domain.Posts;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class PostsSaveRequestDto {
-    private String title;
-    private String content;
-    private String author;
-
-    @Builder
-    public PostsSaveRequestDto(String title, String content, String author) {
-        this.title = title;
-        this.content = content;
-        this.author = author;
-    }
-
-    public Posts toEntity() {
-        return Posts.builder()
-                .title(title)
-                .content(content)
-                .author(author)
-                .build();
-    }
-}
