--- src/main/java/com/simso/dto/post/PostsUpdateRequestDto.java
@@ -1,19 +1,0 @@
-package com.simso.dto.post;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class PostsUpdateRequestDto {
-    private String title;
-    private String content;
-
-
-    @Builder
-    public PostsUpdateRequestDto(String title, String content){
-        this.title = title;
-        this.content = content;
-    }
-}
