--- src/main/java/com/simso/exception/ExceptionAdvice.java
@@ -1,24 +1,0 @@
-package com.simso.exception;
-
-import com.simso.global.common.response.CommonResult;
-import com.simso.global.common.response.ResponseService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import javax.servlet.http.HttpServlet;
-
-@RestControllerAdvice
-@RequiredArgsConstructor
-public class ExceptionAdvice {
-    private final ResponseService responseService;
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR) //500
-    protected CommonResult defaultException(HttpServlet request, Exception e) {
-
-        return responseService.getFailResult();
-    }
-}
