+++ src/main/java/com/simso/global/exception/ExceptionAdvice.java
@@ -1,0 +1,20 @@
+package com.simso.global.exception;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.time.LocalDateTime;
+
+@ControllerAdvice
+@RequiredArgsConstructor
+public class ExceptionAdvice {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    protected ResponseEntity<?> handleIllegalArgumentException(IllegalArgumentException e) {
+        return new ResponseEntity<>(new ErrorResponse(LocalDateTime.now()), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+}
