--- src/main/java/com/simso/repository/PostsRepository.java
@@ -1,13 +1,0 @@
-package com.simso.repository;
-
-import com.simso.domain.Posts;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-
-public interface PostsRepository extends JpaRepository <Posts, Long> {
-
-    @Query("SELECT p FROM Posts p ORDER BY p.id DESC")
-    List<Posts> findAllDesc();
-}
