--- src/main/java/com/simso/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.simso.repository;
-
-import com.simso.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-   Optional<User> findByUsername(String username);
-
-   Optional<User> findByEmail(String email);
-}
