--- src/test/java/com/simso/controller/api/PostsApiControllerTest.java
@@ -1,133 +1,0 @@
-package com.simso.controller.api;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.simso.domain.Posts;
-import com.simso.dto.post.PostsSaveRequestDto;
-import com.simso.dto.post.PostsUpdateRequestDto;
-import com.simso.repository.PostsRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.context.WebApplicationContext;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class PostsApiControllerTest {
-    @LocalServerPort //주입되는 port
-    private int port;
-
-    @Autowired
-    private WebApplicationContext context;
-
-    private MockMvc mvc;
-
-    @Autowired
-    private TestRestTemplate testRestTemplate;
-
-    @Autowired
-    private PostsRepository postsRepository;
-
-    @Autowired
-    public void tearDown() throws Exception {
-        postsRepository.deleteAll();
-    }
-
-    @BeforeEach
-    public  void setup() {
-        mvc = MockMvcBuilders
-                .webAppContextSetup(context)
-                .apply(springSecurity())
-                .build();
-    }
-
-    @Test
-    @DisplayName("등록된다")
-    @WithMockUser(roles = "USER")
-    public void register() throws Exception {
-        //given
-        String title = "title";
-        String content = "content";
-        String author = "author";
-        PostsSaveRequestDto requestDto = PostsSaveRequestDto.builder()
-                .title(title)
-                .content(content)
-                .author(author)
-                .build();
-
-
-        String url = "http://localhost:" + port + "/api/v1/posts/";
-
-        //when
-        mvc.perform(post(url)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(new ObjectMapper().writeValueAsString(requestDto)))
-                .andExpect(status().isOk());
-
-
-        //then
-        List<Posts> all = postsRepository.findAll();
-        assertThat(all.get(0).getTitle()).isEqualTo(title);
-        assertThat(all.get(0).getContent()).isEqualTo(content);
-        assertThat(all.get(0).getAuthor()).isEqualTo(author);
-
-    }
-
-    @Test
-    @DisplayName("수정된다")
-    @Transactional
-//    @Rollback(value = false)
-    @WithMockUser(roles = "USER")
-    public void update() throws Exception {
-        //given
-        String title = "title update";
-        String content = "content update";
-        String author = "author";
-
-        Posts saveUser = postsRepository.save(Posts.builder()
-                .title(title)
-                .content(content)
-                .author(author)
-                .build());
-
-
-        Long updateId = saveUser.getId();
-        String expectedTitle = "title update ex";
-        String expectedContent = "content update ex";
-
-        PostsUpdateRequestDto requestDto = PostsUpdateRequestDto.builder()
-                .title(expectedTitle)
-                .content(expectedContent)
-                .build();
-
-        String url = "http://localhost:" + port + "/api/v1/posts/" + updateId;
-
-
-        //when
-        mvc.perform(put(url)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(new ObjectMapper().writeValueAsString(requestDto)))
-                .andExpect(status().isOk());
-
-        //then
-
-        List<Posts> all = postsRepository.findAll();
-        assertThat(all.get(0).getTitle()).isEqualTo(expectedTitle);
-        assertThat(all.get(0).getContent()).isEqualTo(expectedContent);
-
-    }
-}
