--- src/test/java/com/simso/controller/api/RoadmapApiControllerTest.java
@@ -1,10 +1,0 @@
-package com.simso.controller.api;
-
-import static org.junit.jupiter.api.Assertions.*;
-/*
-*  MVC 테스트
-*  컨트롤러가 예상대로 동작하기 위한 테스트
-* */
-class RoadmapApiControllerTest {
-
-}
