+++ src/test/java/com/simso/global/controller/api/RoadmapApiControllerTest.java
@@ -1,0 +1,9 @@
+package com.simso.global.controller.api;
+
+/*
+*  MVC 테스트
+*  컨트롤러가 예상대로 동작하기 위한 테스트
+* */
+class RoadmapApiControllerTest {
+
+}
