+++ src/test/java/com/simso/repository/UserRepositoryTest.java
@@ -1,12 +1,12 @@
 package com.simso.repository;
 
-import com.simso.domain.Role;
-import com.simso.domain.User;
+import com.simso.domain.user.entity.Role;
+import com.simso.domain.user.entity.User;
+import com.simso.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
